+++ tests/AccountTest.php
@@ -35,6 +35,28 @@ class AccountTest extends TestCase
      * @test
      * @throws ApiException
      */
+    public function itGetsTheBalanceForATestAlias()
+    {
+        $data = $this->arionum->getBalanceByAlias('PXGAMER');
+        $this->assertInternalType('string', $data);
+        $this->assertTrue(is_numeric($data));
+    }
+
+    /**
+     * @test
+     * @throws ApiException
+     */
+    public function itGetsTheBalanceForATestPublicKey()
+    {
+        $data = $this->arionum->getBalanceByPublicKey(self::TEST_PUBLIC_KEY);
+        $this->assertInternalType('string', $data);
+        $this->assertTrue(is_numeric($data));
+    }
+
+    /**
+     * @test
+     * @throws ApiException
+     */
     public function itGetsThePendingBalanceForATestAddress()
     {
         $data = $this->arionum->getPendingBalance(self::TEST_ADDRESS);
@@ -57,10 +79,32 @@ class AccountTest extends TestCase
      * @test
      * @throws ApiException
      */
+    public function itGetsTheTransactionsForATestPublicKey()
+    {
+        $data = $this->arionum->getTransactionsByPublicKey(self::TEST_PUBLIC_KEY);
+        $this->assertInternalType('array', $data);
+        $this->assertNotEmpty($data);
+    }
+
+    /**
+     * @test
+     * @throws ApiException
+     */
     public function itGetsTheAliasForASpecificAddress()
     {
         $data = $this->arionum->getAlias(self::TEST_ADDRESS);
         $this->assertInternalType('string', $data);
         $this->assertEquals('PXGAMER', $data);
+    }
+
+    /**
+     * @test
+     * @throws ApiException
+     */
+    public function itChecksThatTheAddressIsValid()
+    {
+        $data = $this->arionum->checkAddress(self::TEST_ADDRESS);
+        $this->assertInternalType('bool', $data);
+        $this->assertTrue($data);
     }
 }